requirements.txt
# 2025/04/01
langchain_openai
langchain_community
notion_client
python-dotenv
openai
faiss-cpu